Check java version at most once (#27)
