+++ modelator/src/jar.rs
@@ -1,6 +1,12 @@
 use crate::error::Error;
 use std::collections::HashSet;
 use std::path::{Path, PathBuf};
+use std::process::Command;
+
+// The minimum java version supported by Apalache is Java 8:
+// https://apalache.informal.systems/docs/apalache/system-reqs.html
+// TLC doesn't seem to have such requirement.
+const MIN_JAVA_VERSION: usize = 8;
 
 pub const TLA_JAR: &str = "tla2tools-v1.8.0.jar";
 pub const TLA_JAR_BYTES: &[u8] = include_bytes!("../jars/tla2tools-v1.8.0.jar");
@@ -63,10 +69,17 @@ pub(crate) fn write_jars<P: AsRef<Path>>(modelator_dir: P) -> Result<(), Error> 
     // get all existing jars
     let existing_jars = existing_jars(&modelator_dir)?;
     // download all jars that do not exist yet
+    let mut check_java = false;
     for jar in Jar::all() {
         if !existing_jars.contains(&jar) {
             jar.write(&modelator_dir)?;
+            check_java = true;
         }
+    }
+    // if we wrote the jar(s) for the first time, check that we have a
+    // compatible version of java
+    if check_java {
+        check_java_version()?;
     }
     Ok(())
 }
@@ -85,4 +98,50 @@ fn existing_jars<P: AsRef<Path>>(modelator_dir: P) -> Result<HashSet<Jar>, Error
         "[modelator] at most 3 jar files should have been downloaded"
     );
     Ok(existing_jars)
+}
+
+fn check_java_version() -> Result<(), Error> {
+    let mut cmd = Command::new("java");
+    cmd.arg("-XshowSettings:all").arg("--version");
+    // show command being run
+    tracing::debug!("{}", crate::util::cmd_show(&cmd));
+
+    match cmd.output() {
+        Ok(output) => {
+            let stderr = crate::util::cmd_output_to_string(&output.stderr);
+            let mut versions: Vec<_> = stderr
+                .lines()
+                .filter(|line| line.trim().starts_with("java.specification.version ="))
+                .collect();
+            tracing::debug!("java version {:?}", versions);
+
+            assert_eq!(
+                versions.len(),
+                1,
+                "[modelator] expected at most one 'java.specification.version'"
+            );
+
+            // it's okay to unwrap as we have already asserted that the version exists
+            let version_parts: Vec<_> = versions.pop().unwrap().split('=').collect();
+            assert_eq!(
+                version_parts.len(),
+                2,
+                "[modelator] unexpected 'java.specification.version' format"
+            );
+            let version_number: usize = version_parts[1]
+                .trim()
+                .parse()
+                .expect("[modelator] unexpected 'java.specification.version' format");
+
+            if version_number < MIN_JAVA_VERSION {
+                Err(Error::MinimumJavaVersion(version_number, MIN_JAVA_VERSION))
+            } else {
+                Ok(())
+            }
+        }
+        Err(err) => {
+            tracing::debug!("error checking Java version: {}", err);
+            Err(Error::MissingJava)
+        }
+    }
 }
