+++ modelator/src/module/apalache/mod.rs
@@ -24,9 +24,6 @@ impl Apalache {
             options
         );
 
-        // check java and its version
-        crate::util::check_java_version()?;
-
         // check that the tla file and tla cfg file exist
         tla_file.check_existence()?;
         tla_config_file.check_existence()?;
@@ -62,9 +59,6 @@ impl Apalache {
     // #[modelator::method]
     pub fn parse(tla_file: TlaFile, options: &Options) -> Result<TlaFile, Error> {
         tracing::debug!("Apalache::parse {} {:?}", tla_file, options);
-
-        // check java and its version
-        crate::util::check_java_version()?;
 
         // check that the tla file and tla cfg file exist
         tla_file.check_existence()?;
