+++ modelator/src/util.rs
@@ -3,11 +3,6 @@ use std::collections::HashSet;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 
-// The minimum java version supported by Apalache is Java 8:
-// https://apalache.informal.systems/docs/apalache/system-reqs.html
-// TLC doesn't seem to have such requirement.
-const MIN_JAVA_VERSION: usize = 8;
-
 pub(crate) fn cmd_output_to_string(output: &[u8]) -> String {
     String::from_utf8_lossy(output).to_string()
 }
@@ -48,50 +43,4 @@ pub(crate) fn read_dir<P: AsRef<Path>>(path: P) -> Result<HashSet<String>, Error
         assert!(file_names.insert(file_name));
     }
     Ok(file_names)
-}
-
-pub(crate) fn check_java_version() -> Result<(), Error> {
-    let mut cmd = Command::new("java");
-    cmd.arg("-XshowSettings:all").arg("--version");
-    // show command being run
-    tracing::debug!("{}", cmd_show(&cmd));
-
-    match cmd.output() {
-        Ok(output) => {
-            let stderr = cmd_output_to_string(&output.stderr);
-            let mut versions: Vec<_> = stderr
-                .lines()
-                .filter(|line| line.trim().starts_with("java.specification.version ="))
-                .collect();
-            tracing::debug!("java version {:?}", versions);
-
-            assert_eq!(
-                versions.len(),
-                1,
-                "[modelator] expected at most one 'java.specification.version'"
-            );
-
-            // it's okay to unwrap as we have already asserted that the version exists
-            let version_parts: Vec<_> = versions.pop().unwrap().split('=').collect();
-            assert_eq!(
-                version_parts.len(),
-                2,
-                "[modelator] unexpected 'java.specification.version' format"
-            );
-            let version_number: usize = version_parts[1]
-                .trim()
-                .parse()
-                .expect("[modelator] unexpected 'java.specification.version' format");
-
-            if version_number < MIN_JAVA_VERSION {
-                Err(Error::MinimumJavaVersion(version_number, MIN_JAVA_VERSION))
-            } else {
-                Ok(())
-            }
-        }
-        Err(err) => {
-            tracing::debug!("error checking Java version: {}", err);
-            Err(Error::MissingJava)
-        }
-    }
 }
